backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2